Full prescision for X and DRAM for all analogue
Instead of slicing the upper bits of the output of the cross correlation
accumulator, rather keep all bits. This means two snap blocks are
required as datarate is 256 bps per channel and a snap can only handle 128.

Instead of a multiplexed snap to only one analogue channel, snap all
analogue channels in a DRAM snap